// src/app/home/model/home-nav-item.ts
export class HomeNavItem {
    id: number;
    displayName: string;
    menuName: string;
    icon: string;
    notice: string;
    linkAddress: string;
    isDisplay: boolean;
    groupName: string;
    groupDisplayName: string;
    isActive?: string;
    // constructor( values: Object = {} ) {
    //     Object.assign( this , values );
    // }
}